Refactored and optimized code, started work on language support Changed all literal string references to calls to a resource bundle to support different languages. Refactored all the serialization functions to remove duplicate code and remove side effects from the functions actually handling serialization. Moved all the serialization functions to SerializationUtils instead of being in FileSystemUtils . Rewrote unit tests to be better isolated and more consistent. Added additional unit tests to cover more of the project.

+++ src/main/java/com/cameronbarnes/mercury/util/FileSystemUtils.java
@@ -4,9 +4,6 @@ import com.cameronbarnes.mercury.core.Main;
 import com.cameronbarnes.mercury.core.Options;
 import com.cameronbarnes.mercury.core.SavedOngoing;
 import com.cameronbarnes.mercury.stock.Bin;
-import com.cameronbarnes.mercury.stock.Part;
-import com.google.gson.*;
-import com.google.gson.reflect.TypeToken;
 import org.apache.commons.lang3.RandomStringUtils;
 
 import javax.swing.*;
@@ -14,15 +11,14 @@ import javax.swing.filechooser.FileNameExtensionFilter;
 import java.io.File;
 import java.io.FileWriter;
 import java.io.IOException;
-import java.lang.reflect.Type;
 import java.nio.file.Files;
-import java.nio.file.NoSuchFileException;
 import java.nio.file.Path;
 import java.time.Instant;
 import java.util.*;
 
 public final class FileSystemUtils {
 	
+	@SuppressWarnings("ResultOfMethodCallIgnored")
 	public static void createProjectDirs() {
 		
 		Options.IMPORT_FOLDER.mkdirs();
@@ -34,16 +30,6 @@ public final class FileSystemUtils {
 	public static String getLastSubstring(String str, String split) {
 		int num = str.lastIndexOf(split);
 		return num == -1 ? str : str.substring(++num);
-	}
-	
-	public static void clearImportFolder() {
-		
-		Arrays.stream(Objects.requireNonNull(Options.IMPORT_FOLDER.listFiles())).dropWhile(File::isDirectory).forEach(File::delete);
-	}
-	
-	public static void clearProcessFolder() {
-		
-		Arrays.stream(Objects.requireNonNull(Options.PROCESS_FOLDER.listFiles())).dropWhile(File::isDirectory).forEach(File::delete);
 	}
 	
 	public static void moveAllFromProcessToIngest() {
@@ -88,11 +74,15 @@ public final class FileSystemUtils {
 			}
 			if (file.isDirectory())
 				return Optional.of(new File(file.getAbsolutePath() + File.separator + "Cycle Count " + Date.from(Instant.now()).toString().replace(":", "-")+ ".xlsx"));
-			if (getLastSubstring(file.getName(), ".").equalsIgnoreCase("xlsx")) {
-				validFile = true;
+			if (!file.getName().contains(".")) { // If the name entered by the user has no extension we'll add .xlsx to make it valid
+				return Optional.of(new File(file.getAbsolutePath() + ".xlsx"));
 			} else {
-				JOptionPane.showMessageDialog(null, "File name must end with .xlsx");
-				chooser.setSelectedFile(new File("Cycle Count " + Date.from(Instant.now()).toString().replace(":", "-")+ ".xlsx"));
+				if (getLastSubstring(file.getName(), ".").equalsIgnoreCase("xlsx")) {
+					validFile = true;
+				} else {
+					JOptionPane.showMessageDialog(null, "File name must end with .xlsx");
+					chooser.setSelectedFile(new File("Cycle Count " + Date.from(Instant.now()).toString().replace(":", "-")+ ".xlsx"));
+				}
 			}
 		}
 		
@@ -100,111 +90,75 @@ public final class FileSystemUtils {
 		
 	}
 	
-	public static File getDirectoryWithFileChooser() { // Change this to use a Optional with Optional.ofNullable
+	public static Optional<File> getDirectoryWithFileChooser() {
 		
 		JFileChooser chooser = new JFileChooser(Options.IMPORT_FOLDER);
 		chooser.setFileSelectionMode(JFileChooser.DIRECTORIES_ONLY);
 		
 		if (chooser.showOpenDialog(null) == JFileChooser.APPROVE_OPTION) {
-			return chooser.getSelectedFile();
+			return Optional.ofNullable(chooser.getSelectedFile());
 		}
-		else return Options.IMPORT_FOLDER;
+		else return Optional.empty();
 		
 	}
 	
-	public static void writeOptions(Options options) {
-		writeOptions(options, new File("options.json"));
-		writeVersion(new File("version.json"));
-	}
-	
-	public static void writeOptions(Options options, File file) {
+	public static void writeStringToFile(String str, File file) {
 		
 		try {
 			file.createNewFile();
-		}
-		catch (IOException e) {
-			e.printStackTrace();
+		} catch (IOException e) {
+			e.printStackTrace(); // TODO probably handle this with the HomeAPI
 		}
 		
 		try (FileWriter writer = new FileWriter(file)) {
 			
-			Gson gson = new Gson();
-			writer.write(gson.toJson(options));
+			writer.write(str);
 			writer.flush();
 			
 		}
 		catch (IOException e) {
-			throw new RuntimeException(e);
+			e.printStackTrace(); // TODO probably handle this with the HomeAPI
 		}
 		
 	}
 	
-	public static void writeVersion(File file) {
+	public static Optional<String> readStringFromFile(File file) {
 		
-		try {
-			file.createNewFile();
-		}
-		catch (IOException e) {
-			e.printStackTrace();
-		}
-		
-		try (FileWriter writer = new FileWriter(file)) {
+		if (file.exists() && !file.isDirectory()) {
 			
-			Gson gson = new Gson();
-			writer.write(gson.toJson(Main.VERSION));
-			writer.flush();
+			try {
+				String str = new String(Files.readAllBytes(file.toPath()));
+				return str.isBlank() ? Optional.empty() : Optional.of(str);
+			}
+			catch (IOException e) {
+				e.printStackTrace(); // TODO probably handle this with the HomeAPI
+			}
 			
-		}
-		catch (IOException e) {
-			e.printStackTrace();
-		}
-		
-	}
-	
-	public static Optional<Main.Version> readVersion () {
-		return readVersion(new File("version.json"));
-	}
-	
-	public static Optional<Main.Version> readVersion(File file) {
-		
-		try {
-			
-			String value = new String(Files.readAllBytes(Path.of(file.getAbsolutePath())));
-			GsonBuilder builder = new GsonBuilder();
-			builder.registerTypeAdapter(Main.Version.class, new VersionDeserializer());
-			Gson gson = builder.create();
-			return Optional.ofNullable(gson.fromJson(value, Main.Version.class));
-			
-		}
-		catch (NoSuchFileException ignored) {}
-		catch (IOException e) {
-			e.printStackTrace();
 		}
 		
 		return Optional.empty();
 		
 	}
 	
-	public static Options readOptions() {
-		return readOptions(new File("options.json"));
+	public static void writeOptions(Options options) {
+		writeOptions(options, new File("options.json"), new File("version.json"));
 	}
 	
-	public static Options readOptions(File in) {
-		
-		try {
-			
-			String file = new String(Files.readAllBytes(in.toPath()));
-			Gson gson = new Gson();
-			Options options = gson.fromJson(file, Options.class);
-			options.generateFont();
-			options.ensureAllNewPropertiesArePresent();
-			return options;
-			
-		}
-		catch (IOException ignored) {} // This is fine because then we can just return a new options object
-		
-		return new Options();
-		
+	public static void writeOptions(Options options, File optFile, File versionFile) {
+		writeStringToFile(SerializationUtils.serializeOptions(options), optFile);
+		writeStringToFile(SerializationUtils.serializeVersion(Main.VERSION), versionFile);
+	}
+	
+	public static Options readOptions() {
+		return readOptions(new File("options.json"), new File("version.json"));
+	}
+	
+	public static Options readOptions(File optionsFile, File versionFile) {
+	
+		Options options = readStringFromFile(optionsFile).flatMap(SerializationUtils::deserializeOptions).orElse(new Options());
+		readStringFromFile(versionFile).flatMap(SerializationUtils::deserializeVersion).ifPresent(options::setVersion);
+		return options;
+	
 	}
 	
 	public static void saveOngoing(List<Bin> bins) {
@@ -220,23 +174,8 @@ public final class FileSystemUtils {
 		File binsDir = new File(outDir.getAbsolutePath() + File.separator + "bins");
 		binsDir.mkdirs();
 		
-		// Write out all the parts from bins into json files
-		Gson gson =  new GsonBuilder().serializeNulls().create();
-		bins.forEach(bin -> {
-			File out = new File(binsDir.getAbsolutePath() + File.separator + bin.getBinNum());
-			try {
-				out.createNewFile();
-				try (FileWriter writer = new FileWriter(out)) {
-					
-					writer.write(gson.toJson(new BinData(bin.getBinNum(), bin.getWarehouse(), gson.toJson(bin.getParts()))));
-					writer.flush();
-					
-				}
-			}
-			catch (IOException e) {
-				throw new RuntimeException(e);
-			}
-		});
+		//Write out all the parts from bins into json files
+		bins.forEach(bin -> writeStringToFile(SerializationUtils.serializeBin(bin), new File(binsDir.getAbsolutePath() + File.separator + bin.getBinNum())));
 		
 		// Move all the stockstatus files into a folder with the save
 		Arrays.stream(Objects.requireNonNull(Options.PROCESS_FOLDER.listFiles())).dropWhile(File::isDirectory).forEach(file -> {
@@ -244,32 +183,14 @@ public final class FileSystemUtils {
 				Files.move(file.toPath(), Path.of(stockstatusDir.getAbsolutePath() + File.separator + file.getName()));
 			}
 			catch (IOException e) {
-				e.printStackTrace();
+				e.printStackTrace(); // TODO handle this with the HomeAPI
 			}
 		});
 		
 	}
 	
 	public static boolean hasSavedSessions() {
-		
 		return Objects.requireNonNull(Options.SAVED_ONGOING_FOLDER.listFiles()).length > 0;
-		
-	}
-	
-	private static Optional<Bin> loadBinFromFile(File file) {
-		
-		GsonBuilder builder = new GsonBuilder();
-		builder.registerTypeAdapter(BinData.class, new BinDataDeserializer());
-		Gson gson = builder.create();
-		try {
-			String json = new String(Files.readAllBytes(file.toPath()));
-			BinData binData = gson.fromJson(json, BinData.class);
-			return Optional.of(binData.toBin());
-		}
-		catch (IOException e) {
-			return Optional.empty();
-		}
-		
 	}
 	
 	public static List<SavedOngoing> getSavedSessions() {
@@ -283,20 +204,19 @@ public final class FileSystemUtils {
 	
 	public static Optional<SavedOngoing> getSavedSessionFromDir(File dir) {
 		
-		if (Arrays.stream(Objects.requireNonNull(dir.listFiles())).anyMatch(file -> file.getName().equals("stockstatus")) &&
+		// Require tha the provided directory exists,  contains a stockstatus directory, and contains a bins directory
+		if (dir.exists() && Arrays.stream(Objects.requireNonNull(dir.listFiles())).anyMatch(file -> file.getName().equals("stockstatus")) &&
 				    Arrays.stream(Objects.requireNonNull(dir.listFiles())).anyMatch(file -> file.getName().equals("bins"))) {
 			
 			ArrayList<Bin> bins = new ArrayList<>();
 			
+			// Get all files in the bins subdirectory, exclude directories, read the contents of each file as a string, deserialize from string to bin, add the valid bins to the array
 			Arrays.stream(Objects.requireNonNull(new File(dir.getAbsolutePath() + "/bins").listFiles()))
-			      .dropWhile(File::isDirectory).forEach(file -> loadBinFromFile(file).ifPresent(bins::add));
+			      .dropWhile(File::isDirectory).map(FileSystemUtils::readStringFromFile)
+			      .filter(Optional::isPresent).map(Optional::get).map(SerializationUtils::deserializeBin).filter(Optional::isPresent).map(Optional::get).forEach(bins::add);
 			
 			if (!bins.isEmpty())
-				return Optional.of(new SavedOngoing(
-						dir,
-						(int) Arrays.stream(Objects.requireNonNull(new File(dir.getAbsolutePath() + File.separator + "stockstatus").listFiles())).dropWhile(File::isDirectory).count(),
-						bins)
-				);
+				return Optional.of(new SavedOngoing(dir, bins));
 			
 		}
 		
@@ -306,7 +226,7 @@ public final class FileSystemUtils {
 	
 	public static void moveFilesFromSavedSession(SavedOngoing savedOngoing) {
 		
-		// We'll move all the stockstatus files from the saveed data to the process folder
+		// We'll move all the stockstatus files from the saved data to the process folder
 		Arrays.stream(Objects.requireNonNull(new File(savedOngoing.getSaveDir().getAbsolutePath() + File.separator + "stockstatus").listFiles())).dropWhile(File::isDirectory).forEach(file -> {
 			try {
 				Files.move(file.toPath(), Path.of(Options.PROCESS_FOLDER.getPath() + File.separator + file.getName()));
@@ -320,44 +240,16 @@ public final class FileSystemUtils {
 		
 	}
 	
+	@SuppressWarnings("ResultOfMethodCallIgnored")
 	public static void deleteDir(File dir) {
 		
 		if (!dir.exists())
 			return;
+		if (!dir.isDirectory()) // This shouldn't ever come up, but I'd rather this not fail silently if it does
+			throw new RuntimeException("Provided file must be a directory"); // TODO handle this with the HomeAPI
 		Arrays.stream(Objects.requireNonNull(dir.listFiles())).filter(File::isDirectory).forEach(FileSystemUtils::deleteDir);
 		Arrays.stream(Objects.requireNonNull(dir.listFiles())).forEach(File::delete);
 		dir.delete();
-		
-	}
-	
-	private record BinData(String BinNumber, String WareHouse, String Parts) {
-		
-		Bin toBin() {
-			
-			Gson gson = new Gson();
-			Type type = new TypeToken<ArrayList<Part>>(){}.getType();
-			ArrayList<Part> parts = gson.fromJson(Parts, type);
-			return new Bin(BinNumber, WareHouse, parts);
-			
-		}
-		
-	}
-	
-	private static final class BinDataDeserializer implements JsonDeserializer<BinData> {
-		@Override
-		public BinData deserialize(JsonElement jsonElement, Type type, JsonDeserializationContext jsonDeserializationContext) throws JsonParseException {
-			JsonObject jsonObject = jsonElement.getAsJsonObject();
-			return new BinData(jsonObject.get("BinNumber").getAsString(), jsonObject.get("WareHouse").getAsString(), jsonObject.get("Parts").getAsString());
-		}
-		
-	}
-	
-	private static final class VersionDeserializer implements JsonDeserializer<Main.Version> {
-		@Override
-		public Main.Version deserialize(JsonElement json, Type typeOfT, JsonDeserializationContext context) throws JsonParseException {
-			JsonObject jsonObject = json.getAsJsonObject();
-			return new Main.Version(jsonObject.get("major").getAsInt(), jsonObject.get("minor").getAsInt(), jsonObject.get("patch").getAsInt(), Main.ReleaseType.valueOf(jsonObject.get("label").getAsString()));
-		}
 		
 	}
